requirements.txt
fastapi
uvicorn[standard]
python-multipart
pydantic
pydantic-settings
streamlit
opencv-python
numpy
Pillow
pandas
easyocr
ultralytics
PyYAML
kaggle
